requirements.txt
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
pytz==2019.1
psycopg2==2.8.2
django-cors-headers==3.0.1
requests==2.22.0